// apps/web/src/data/subgraph/sdk.generated.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  BigDecimal: any;
  BigInt: any;
  Bytes: any;
  Int8: any;
};

export type Auction = {
  __typename?: 'Auction';
  bidCount: Scalars['Int'];
  bids?: Maybe<Array<AuctionBid>>;
  dao: Dao;
  endTime: Scalars['BigInt'];
  extended: Scalars['Boolean'];
  firstBidTime?: Maybe<Scalars['BigInt']>;
  highestBid?: Maybe<AuctionBid>;
  id: Scalars['ID'];
  settled: Scalars['Boolean'];
  startTime: Scalars['BigInt'];
  token: Token;
  winningBid?: Maybe<AuctionBid>;
};


export type AuctionBidsArgs = {
  first?: InputMaybe<Scalars['Int']>;
  orderBy?: InputMaybe<AuctionBid_OrderBy>;
  orderDirection?: InputMaybe<OrderDirection>;
  skip?: InputMaybe<Scalars['Int']>;
  where?: InputMaybe<AuctionBid_Filter>;
};

export type AuctionBid = {
  __typename?: 'AuctionBid';
  amount: Scalars['BigInt'];
  auction: Auction;
  bidTime: Scalars['BigInt'];
  bidder: Scalars['Bytes'];
  id: Scalars['ID'];
};
// ...
export type AuctionConfig = {
  __typename?: 'AuctionConfig';
  duration: Scalars['BigInt'];
  id: Scalars['ID'];
  minimumBidIncrement: Scalars['BigInt'];
  reservePrice: Scalars['BigInt'];
  timeBuffer: Scalars['BigInt'];
};
// ...
export enum Auction_OrderBy {
  BidCount = 'bidCount',
  Bids = 'bids',
  Dao = 'dao',
  DaoAuctionAddress = 'dao__auctionAddress',
  DaoContractImage = 'dao__contractImage',
  DaoDescription = 'dao__description',
  DaoGovernorAddress = 'dao__governorAddress',
  DaoId = 'dao__id',
  DaoMetadataAddress = 'dao__metadataAddress',
  DaoName = 'dao__name',
  DaoOwnerCount = 'dao__ownerCount',
  DaoProjectUri = 'dao__projectURI',
  DaoProposalCount = 'dao__proposalCount',
  DaoSymbol = 'dao__symbol',
  DaoTokenAddress = 'dao__tokenAddress',
  DaoTotalAuctionSales = 'dao__totalAuctionSales',
  DaoTotalSupply = 'dao__totalSupply',
  DaoTreasuryAddress = 'dao__treasuryAddress',
  EndTime = 'endTime',
  Extended = 'extended',
  FirstBidTime = 'firstBidTime',
  HighestBid = 'highestBid',
  HighestBidAmount = 'highestBid__amount',
  HighestBidBidTime = 'highestBid__bidTime',
  HighestBidBidder = 'highestBid__bidder',
  HighestBidId = 'highestBid__id',
  Id = 'id',
  Settled = 'settled',
  StartTime = 'startTime',
  Token = 'token',
  TokenId = 'token__id',
  TokenImage = 'token__image',
  TokenMintedAt = 'token__mintedAt',
  TokenName = 'token__name',
  TokenOwner = 'token__owner',
  TokenTokenContract = 'token__tokenContract',
  TokenTokenId = 'token__tokenId',
  WinningBid = 'winningBid',
  WinningBidAmount = 'winningBid__amount',
  WinningBidBidTime = 'winningBid__bidTime',
  WinningBidBidder = 'winningBid__bidder',
  WinningBidId = 'winningBid__id'
}

export type BlockChangedFilter = {
  number_gte: Scalars['Int'];
};

export type Block_Height = {
  hash?: InputMaybe<Scalars['Bytes']>;
  number?: InputMaybe<Scalars['Int']>;
  number_gte?: InputMaybe<Scalars['Int']>;
};

export type Dao = {
  __typename?: 'DAO';
  auctionAddress: Scalars['Bytes'];
  auctionConfig: AuctionConfig;
  auctions: Array<Auction>;
  contractImage: Scalars['String'];
  currentAuction?: Maybe<Auction>;
  description: Scalars['String'];
  governorAddress: Scalars['Bytes'];
  id: Scalars['ID'];
  metadataAddress: Scalars['Bytes'];
  name: Scalars['String'];
  ownerCount: Scalars['Int'];
  owners: Array<DaoTokenOwner>;
  projectURI: Scalars['String'];
  proposalCount: Scalars['Int'];
  proposals: Array<Proposal>;
  symbol: Scalars['String'];
  tokenAddress: Scalars['Bytes'];
  tokens: Array<Token>;
  totalAuctionSales: Scalars['BigInt'];
  totalSupply: Scalars['Int'];
  treasuryAddress: Scalars['Bytes'];
};
// ...
export type DaoTokenOwner = {
  __typename?: 'DAOTokenOwner';
  dao: Dao;
  daoTokenCount: Scalars['Int'];
  daoTokens: Array<Token>;
  id: Scalars['ID'];
  owner: Scalars['Bytes'];
};
// ...
export enum OrderDirection {
  Asc = 'asc',
  Desc = 'desc'
}

export type Proposal = {
  __typename?: 'Proposal';
  abstainVotes: Scalars['Int'];
  againstVotes: Scalars['Int'];
  calldatas?: Maybe<Scalars['String']>;
  canceled: Scalars['Boolean'];
  dao: Dao;
  description?: Maybe<Scalars['String']>;
  descriptionHash: Scalars['Bytes'];
  executableFrom?: Maybe<Scalars['BigInt']>;
  executed: Scalars['Boolean'];
  expiresAt?: Maybe<Scalars['BigInt']>;
  forVotes: Scalars['Int'];
  id: Scalars['ID'];
  proposalId: Scalars['Bytes'];
  proposalNumber: Scalars['Int'];
  proposalThreshold: Scalars['BigInt'];
  proposer: Scalars['Bytes'];
  queued: Scalars['Boolean'];
  quorumVotes: Scalars['BigInt'];
  snapshotBlockNumber: Scalars['BigInt'];
  targets: Array<Scalars['Bytes']>;
  timeCreated: Scalars['BigInt'];
  title?: Maybe<Scalars['String']>;
  transactionHash: Scalars['Bytes'];
  values: Array<Scalars['BigInt']>;
  vetoed: Scalars['Boolean'];
  voteCount: Scalars['Int'];
  voteEnd: Scalars['BigInt'];
  voteStart: Scalars['BigInt'];
  votes: Array<ProposalVote>;
};
// ...
export type Token = {
  __typename?: 'Token';
  auction?: Maybe<Auction>;
  dao: Dao;
  id: Scalars['ID'];
  image: Scalars['String'];
  mintedAt: Scalars['BigInt'];
  name: Scalars['String'];
  owner: Scalars['Bytes'];
  ownerInfo: DaoTokenOwner;
  tokenContract: Scalars['Bytes'];
  tokenId: Scalars['BigInt'];
};
// ...
export type ExploreDaoFragment = { __typename?: 'Auction', endTime: any, dao: { __typename?: 'DAO', id: string, name: string, tokenAddress: any, contractImage: string, description: string, ownerCount: number, proposalCount: number, totalAuctionSales: any, totalSupply: number, symbol: string, projectURI: string }, highestBid?: { __typename?: 'AuctionBid', amount: any, bidder: any } | null, token: { __typename?: 'Token', name: string, image: string, tokenId: any } };